+++ README.md
@@ -46,7 +46,7 @@ To generate the Graphql schema locally run the following command:
 The generated schema file will be placed in the following location: `target/sbt-graphql/schema.graphql`. You can copy the contents of this file into `./schema.graphql` and commit the changes to allow the build to pass.
 
 #### Running the tests 
-The repository and route tests now use a docker container for the database. To run the tests locally, you will need to build this database.
+The repository and route tests use a docker container for the database. To run the tests locally, you will need to build this database.
 ```shell
 docker build -f Dockerfile-tests -t tests .
 ```
